extract UI thread test fakes into tests/ui/_fakes module
