+++ tests/ui/__init__.py
@@ -1,0 +1,1 @@
+"""UI test package; declared so test modules can share helpers via relative imports."""
